move location and update pyproject dependencies
