+++ pyproject.toml
@@ -1,12 +1,23 @@
 [project]
 name = "gaphor-autolayout"
 version = "0.1.0"
-description = "Add your description here"
+description = "Auto layout diagrams using ELK"
 readme = "README.md"
 requires-python = ">=3.12"
-dependencies = [
-    "gaphor>=3.0.0",
+
+[project.entry-points."gaphor.services"]
+"ELK_autolayout" = "gaphor-autolayout:AutoLayoutELK"
+
+[tool.pytest.ini_options]
+testpaths = [
+    "tests",
+]
+addopts = [
+    "--import-mode=importlib",
 ]
 
-[tool.poetry.plugins."gaphor.services"]
-"ELK _autolayout" = "gaphor-autolayout.autolayout:ExportPlugin"
+[dependency-groups]
+dev = [
+    "gaphor>=3.0.0",
+    "pytest>=8.3.4",
+]
